demo-system/service2/requirements.txt
Flask==1.1.2
python-json-logger==2.0.1
gunicorn==20.0.4 
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-mysql
opentelemetry-exporter-otlp
opentelemetry-exporter-opencensus
grpcio
opencensus-proto
opentelemetry-api
opentelemetry-sdk
datadog

